+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 0.12 — Unreleased
 
+- Report profit/loss information when printing deal confirmations
 - Renamed `--print-requests` option to `--verbose`
 - The `--currency-code` option to `ig_markets positions create` and `ig_markets orders create` no longer defaults to
   `USD` and must be specified explicitly